last update to remove the vista fixes on d2007 had a bug, as a result wormtail was not shown on the taskbar - fixed now;
also, we now show the main form on the taskbar again, not the application secret window. that also means though that the taskbar item caption will always be the same as the main form caption, and, as it is currently, will not include the full filepath, but only the filename.